simplify type system: convert all timestamps from int64 to int
Minimize number of int types.
Going with just int for now.

+++ alive.go
@@ -9,7 +9,7 @@ type Alive struct {
 	Producer Producer `xml:"product,attr" json:"producer"`
 	// Timestamp in milliseconds since epoch when this message was generated
 	// according to generating system's clock.
-	Timestamp int64 `xml:"timestamp,attr" json:"timestamp"`
+	Timestamp int `xml:"timestamp,attr" json:"timestamp"`
 	// If set to 0 this means the product is up again after downtime, and the
 	// receiving client will have to issue recovery messages against the API to
 	// start receiving any additional messages and get the current state.
