+++ uof.go
@@ -3,5 +3,5 @@ package uof
 type QueueMsg struct {
 	RoutingKey string
 	Body       []byte
-	Timestamp  int64
+	Timestamp  int
 }
